proc_suite

Module lists

Parse comma-separated lists with built-in diagnostics.

Use provided list types directly with Parse

Parse failures automatically emit structured diagnostics via ParseDiagnostic

Use DuplicateCheck::duplicate_check to validate and report duplicates

Structs

BStringList

A wrapper around a Punctuated<LitByteStr, Comma> representing a list of byte string literals provided to parse.

BoundsList

A wrapper around a Punctuated<WherePredicate, Comma> representing a list of trait/lifetime bounds.

ExprList

A wrapper around a Punctuated<Expr, Comma> representing a comma-separated list of expressions.

IdentList

A wrapper around a Punctuated<Ident, Comma> representing a list of identifiers provided to parse.

IntList

A wrapper around a Punctuated<LitInt, Comma> representing a list of integer literals provided to parse.

StmtList

A wrapper around a Punctuated<Stmt, Comma> representing a comma-separated list of Rust statements.

Traits

DuplicateCheck

Validates duplicate entries in list-like parsed arguments.