proc_suite

Macro collect

macro_rules! collect {
    ($($expr:expr),* $(,)?) => { ... };
}

Combines multiple ToTokens expressions into a single proc_macro2::TokenStream.

Each expression can also be a function call whose result implements ToTokens:

let tokens = collect!(
    generate_struct(),
    generate_impl(),
    generate_function(),
);

Conceptually, the macro expands the collected values through a quote! repetition:

let items = vec![expr1, expr2, expr3];
quote! {
    #(#items)*
}

An empty invocation produces an empty proc_macro2::TokenStream:

let tokens = collect!();