proc_suite

Macro error_spaces

macro_rules! error_spaces {
    (
        space: $namespace:expr,
        maintain: $maintain:expr,

        $(
            $(#[$enum_meta:meta])*
            $vis:vis enum $Enum:ident {
                range: $start:literal ..= $end:literal,
                variants: {
                    $(
                        $(#[$variant_meta:meta])*
                        $Var:ident $(=> $desc:literal)?
                    ),+ $(,)?
                }
            }
        )+ $(,)?
    ) => { ... };
    (@check_overlap
        ($E1:ident, $S1:literal, $E1End:literal),
        $( ($E2:ident, $S2:literal, $E2End:literal) ),+
    ) => { ... };
    (@check_overlap ($Enum:ident, $Start:literal, $End:literal)) => { ... };
    (@check_contiguous
        ($E1:ident, $S1:literal, $E1End:literal),
        ($E2:ident, $S2:literal, $E2End:literal)
        $(, $rest:tt)*
    ) => { ... };
    (@check_contiguous ($Enum:ident, $Start:literal, $End:literal)) => { ... };
    (@desc $d:literal) => { ... };
    (@desc) => { ... };
}

Defines multiple macro-system error spaces with contiguous, non-overlapping ranges.

Each error space provides concrete error IDs derived from enum variants, enabling compile-time introspection, debugging, and issue resolution.

Each variant is assigned a sequential error index within its range (in declaration order).

Attaches maintainer + hosted repository metadata via ErrorMaintainers.

space is a string literal used as a short identifier in error output.

Syntax

error_spaces! {
    space: "CNT",                       // Required: short identifier
    maintain: MaintainInfo,             // Required: maintainer metadata

    pub enum CounterErrors {            // Required: error space enum
        range: 1000..=1099,             // Required: unique range
        variants: {                     // Required: variants
            InvalidInput,
            Overflow
        }
    }

    pub enum ParseErrors {
        range: 1100..=1199,             // Must be contiguous
        variants: {
            UnexpectedToken,
            MissingField
        }
    }
}

Example

let err = CounterErrors::InvalidInput;

let id = err.to_error_id();     // [CNT:1000]
let info = err.to_error_info(); // includes desc, location, maintainers

See also

ErrorMaintainers

crate::diagnostics!

crate::error_space!

crate::token_err!

crate::callsite_err!