proc_suite::syns

Trait FirstDocAttribute

pub trait FirstDocAttribute {
    // Required method
    fn first_doc_attribute(&mut self, op: &mut AttributesFirstDoc);
}

Provides the ability to locate the first Rustdoc attribute on a syntax node.

The returned operation records the index of the first matching attribute, if one exists.

Required Methods

fn first_doc_attribute(&mut self, op: &mut AttributesFirstDoc)

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".